Separated mas body
