+++ FrontEnd/src/Pages/ControlPanel/PagesCP/Alumnos/Alumnos.jsx
@@ -8,10 +8,12 @@ import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import { CSVLink } from 'react-csv';
 import Select from 'react-select';
-import Actions from './ActAlumnos';
+import Actions from '../../../../Components/ControlPanel/Alumnos/ActAlumnos';
 import { deleteStudent, getStudents, updateStudent } from '../../../../api/students';
 import { findPeriodos, getPeriodos } from '../../../../api/Periodos';
 import { getClassStudent } from '../../../../api/classStudent';
+import { alumnoVacio } from '../../../../utils/constants';
+import MasInformacionAlumno from '../../../../Components/ControlPanel/Alumnos/MasInformacionAlumno';
 
 export default function Alumnos() {
   const [data, setData] = useState([]);
@@ -56,28 +58,7 @@ export default function Alumnos() {
   const [modalMas, setModalMas] = useState(false);
   const [modalEditar, setModalEditar] = useState(false);
   const [modalEliminar, setModalEliminar] = useState(false);
-  const [consolaSeleccionada, setConsolaSeleccionada] = useState({
-    _id: '',
-    clave_unica_identificacion: '',
-    curp: '',
-    nombre: '',
-    apellido_paterno: '',
-    apellido_materno: '',
-    fecha_de_nacimiento: '',
-    tutor_nombre: '',
-    tutor_apellido_paterno: '',
-    tutor_apellido_materno: '',
-    tutor_correo: '',
-    tutor_num_telefono: '',
-    num_telefono: '',
-    pais: '',
-    estado: '',
-    ciudad: '',
-    colonia: '',
-    codigo_postal: '',
-    escolaridad: '',
-    ultima_escuela: '',
-  });
+  const [consolaSeleccionada, setConsolaSeleccionada] = useState(alumnoVacio);
 
   const handleChange = (e) => {
     const { name, value } = e.target;
@@ -115,8 +96,7 @@ export default function Alumnos() {
     try {
       await updateStudent({
         _id: consolaSeleccionada._id,
-        clave_unica_identificacion:
-                    consolaSeleccionada.clave_unica_identificacion,
+        clave_unica_identificacion: consolaSeleccionada.clave_unica_identificacion,
         curp: consolaSeleccionada.curp,
         nombre: consolaSeleccionada.nombre,
         apellido_paterno: consolaSeleccionada.apellido_paterno,
@@ -154,428 +134,6 @@ export default function Alumnos() {
       console.log(error);
     }
   };
-
-  const bodyMas = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 550,
-        height: 660,
-        backgroundColor: '#fefefd',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-      }}
-    >
-      <h1
-        style={{
-          paddingBottom: '15px',
-          marginTop: '5px',
-          fontFamily: 'arial',
-          width: 500,
-        }}
-        align="center"
-      >
-        Informacion completa de Alumno
-      </h1>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Nombre:
-          {' '}
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.nombre}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Apellido paterno:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.apellido_paterno}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Apellido materno:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.apellido_materno}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Fecha de Nacimiento:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.fecha_de_nacimiento}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Telefono Estudiante:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.num_telefono}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Nombre del Tutor:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.tutor_nombre}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 20px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Apellido paterno del Tutor:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.tutor_apellido_paterno}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 20px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Apellido materno del Tutor:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.tutor_apellido_materno}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Correo del Tutor:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.tutor_correo}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Telefono del Tutor:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.tutor_num_telefono}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Pais:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.pais}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Estado:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.estado}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Ciudad:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.ciudad}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Colonia:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.colonia}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Codigo Postal:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.codigo_postal}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Escolaridad:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.escolaridad}
-        </p>
-      </Box>
-      <Box
-        sx={{
-          display: 'grid',
-          gridAutoFlow: 'row',
-          gridTemplateColumns: 'repeat(2, 1fr)',
-          gridTemplateRows: 'repeat(2, 10px)',
-          gap: 1,
-        }}
-      >
-        <p
-          style={{
-            align: 'justify',
-            fontFamily: 'arial',
-            fontSize: 20,
-            fontWeight: 'bold',
-          }}
-        >
-          Ultima Escuela:
-        </p>
-        <p style={{ align: 'justify', fontFamily: 'arial', fontSize: 20 }}>
-          {consolaSeleccionada.ultima_escuela}
-        </p>
-      </Box>
-    </div>
-  );
 
   const bodyEditar = (
     <div
@@ -956,7 +514,7 @@ export default function Alumnos() {
           }}
         />
         <Modal open={modalMas} onClose={() => abrirCerrarModalMas()}>
-          {bodyMas}
+          <MasInformacionAlumno consolaSeleccionada={consolaSeleccionada}/>
         </Modal>
         <Modal open={modalEditar} onClose={() => abrirCerrarModalEditar()}>
           {bodyEditar}
